requirements.txt
# Database drivers
pymongo>=4.6.0
cassandra-driver>=3.28.0
cassandra-driver[cql]>=3.28.0

# LLM integration
openai>=1.12.0
requests>=2.31.0

# ML/Data Science
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
